# docs/swagger.yaml
definitions:
  models.ADD_TO_BALANCE_BODY:
    properties:
      amount:
        type: number
      id:
        type: integer
      remarks:
        type: string
      transaction_pin:
        type: string
    required:
    - amount
    - id
    - transaction_pin
    type: object
  models.Error:
    properties:
      dev_reason:
        type: string
      message:
        type: string
    type: object
...
  models.HTTP_TRANSACTION_BY_ID_RESPONSE:
    properties:
      data:
        $ref: '#/definitions/models.TRANSACTION-models_USER'
      message:
        type: string
    type: object
...
  models.HTTP_WALLET_RESPONSE:
    properties:
      data:
        $ref: '#/definitions/models.WALLET_REQUEST'
      message:
        type: string
    type: object
  models.TRANSACTION-models_USER:
    properties:
      amount:
        type: number
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      receiver:
        $ref: '#/definitions/models.USER'
      receiver_wallet:
        type: integer
      remarks:
        type: string
      sender:
        $ref: '#/definitions/models.USER'
      sender_wallet:
        type: integer
      status:
        type: string
      updated_at:
        type: string
    required:
    - amount
    - receiver
    - receiver_wallet
    - sender
    - sender_wallet
    - status
    type: object
  models.USER:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_verified:
        type: boolean
      last_name:
        type: string
      password:
        type: string
      phone:
        type: string
      tag:
        type: string
      transaction_pin:
        type: string
      updated_at:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - tag
    - transaction_pin
    type: object
...
  models.WALLET_REQUEST:
    properties:
      balance:
        type: number
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
info:
  contact:
    email: jbayilara@gmail.com
    name: John Ayilara (Jaystar)
    url: https://github.com/Jaystar-Bee
  description: OPEN BANK API
  title: OPEN BANK API
  version: "1.0"
paths:
...
  /wallet:
    get:
      consumes:
      - application/json
      description: Get Wallet
      produces:
      - application/json
      responses:
        "200":
          description: wallet fetched successfully
          schema:
            $ref: '#/definitions/models.HTTP_WALLET_RESPONSE'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Get Wallet
      tags:
      - Wallet
  /wallet/send:
    post:
      consumes:
      - application/json
      description: Send money to another user
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ADD_TO_BALANCE_BODY'
      produces:
      - application/json
      responses:
        "200":
          description: wallet updated successfully
          schema:
            $ref: '#/definitions/models.HTTP_TRANSACTION_BY_ID_RESPONSE'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Send money
      tags:
      - Wallet
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"